test: migrating to jest
