+++ package.json
@@ -20,7 +20,7 @@
     "build": "babel ./src --out-dir ./lib",
     "clean": "del lib .nyc_output",
     "eslint": "eslint ./src",
-    "eslint-fix": "eslint --fix ./src",
+    "eslint-fix": "eslint --fix ./src ./test/*.js",
     "coverage": "nyc --reporter=lcov --reporter=text yarn test",
     "test": "jest --no-cache"
   },
@@ -79,7 +79,6 @@
     "@babel/preset-env": "~7.18.2",
     "@babel/preset-react": "^7.17.12",
     "@babel/register": "~7.17.7",
-    "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^13.3.0",
     "@trendmicro/babel-config": "~1.0.2",
     "codecov": "^3.8.3",
@@ -97,8 +96,8 @@
     "jest": "^28.1.1",
     "jest-environment-jsdom": "^28.1.1",
     "jest-extended": "^2.0.0",
-    "react": "^16.10.0",
-    "react-dom": "^16.10.0",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-i18next": "^11.12.0",
     "sha1": "^1.1.1",
     "text-table": "^0.2.0"
@@ -127,6 +126,6 @@
     ],
     "testPathIgnorePatterns": [],
     "testTimeout": 30000,
-    "testEnvironment": "jsdom"
+    "testEnvironment": "node"
   }
 }
