+++ test/fixtures/custom-separators.js
