+++ test/fixtures/dynamic-keys.js
