+++ test/fixtures/escape-sequences.js
