+++ test/fixtures/i18n/de/resource.json
