+++ test/fixtures/i18n/fr/resource.json
