+++ test/fixtures/i18n/it/resource.json
