+++ test/fixtures/i18n/ja/resource.json
