+++ test/fixtures/modules/handlebars-helper-i18n.hbs
