+++ test/fixtures/multiline-dos.js
