+++ test/fixtures/multiline-unix.js
