+++ test/fixtures/namespace.js
