+++ test/fixtures/trans-defaultValue.jsx
